Split _ssh_client_base. Closes-issue: #9
* add note about context manager for SSH:
  not recommended with fast reconnect
