+++ exec_helpers/_ssh_client_base.py
@@ -22,7 +22,6 @@ import base64
 # noinspection PyCompatibility
 import concurrent.futures
 import copy
-import io  # noqa  # pylint: disable=unused-import
 import logging
 import platform
 import stat
@@ -42,8 +41,9 @@ from exec_helpers import exceptions
 from exec_helpers import exec_result
 from exec_helpers import _log_templates
 from exec_helpers import proc_enums
+from exec_helpers import ssh_auth
 
-__all__ = ('SSHAuth', 'SSHClientBase')
+__all__ = ('SSHClientBase', )
 
 logger = logging.getLogger(__name__)
 logging.getLogger('paramiko').setLevel(logging.WARNING)
@@ -65,249 +65,6 @@ _type_execute_async = typing.Tuple[
 ]
 
 CPYTHON = 'CPython' == platform.python_implementation()
-
-
-class SSHAuth(object):
-    """SSH Authorization object."""
-
-    __slots__ = (
-        '__username', '__password', '__key', '__keys',
-        '__key_filename', '__passphrase'
-    )
-
-    def __init__(
-        self,
-        username=None,  # type: typing.Optional[str]
-        password=None,  # type: typing.Optional[str]
-        key=None,  # type: typing.Optional[paramiko.RSAKey]
-        keys=None,  # type: typing.Optional[typing.Iterable[paramiko.RSAKey]],
-        key_filename=None,  # type: typing.Union[typing.List[str], str, None]
-        passphrase=None,  # type: typing.Optional[str]
-    ):
-        """SSH credentials object.
-
-        Used to authorize SSHClient.
-        Single SSHAuth object is associated with single host:port.
-        Password and key is private, other data is read-only.
-
-        :param username: remote username.
-        :type username: typing.Optional[str]
-        :param password: remote password
-        :type password: typing.Optional[str]
-        :param key: Main connection key
-        :type key: typing.Optional[paramiko.RSAKey]
-        :param keys: Alternate connection keys
-        :type keys: typing.Optional[typing.Iterable[paramiko.RSAKey]]]
-        :param key_filename: filename(s) for additional key files
-        :type key_filename: typing.Union[typing.List[str], str, None]
-        :param passphrase: passphrase for keys. Need, if differs from password
-        :type passphrase: typing.Optional[str]
-
-        .. versionchanged:: 1.0.0
-            added: key_filename, passphrase arguments
-        """
-        self.__username = username
-        self.__password = password
-        self.__key = key
-        self.__keys = [None]
-        if key is not None:
-            # noinspection PyTypeChecker
-            self.__keys.append(key)
-        if keys is not None:
-            for k in keys:
-                if k not in self.__keys:
-                    self.__keys.append(k)
-        self.__key_filename = key_filename
-        self.__passphrase = passphrase
-
-    @property
-    def username(self):  # type: () -> str
-        """Username for auth.
-
-        :rtype: str
-        """
-        return self.__username
-
-    @staticmethod
-    def __get_public_key(key):  # type: () -> typing.Optional[str]
-        """Internal method for get public key from private.
-
-        :type key: paramiko.RSAKey
-        """
-        if key is None:
-            return None
-        return '{0} {1}'.format(key.get_name(), key.get_base64())
-
-    @property
-    def public_key(self):  # type: () -> typing.Optional[str]
-        """public key for stored private key if presents else None.
-
-        :rtype: str
-        """
-        return self.__get_public_key(self.__key)
-
-    @property
-    def key_filename(
-        self
-    ):  # type: () -> typing.Union[typing.List[str], str, None]
-        """Key filename(s).
-
-        .. versionadded:: 1.0.0
-        """
-        return copy.deepcopy(self.__key_filename)
-
-    def enter_password(self, tgt):  # type: (io.StringIO) -> None
-        """Enter password to STDIN.
-
-        Note: required for 'sudo' call
-
-        :param tgt: Target
-        :type tgt: file
-        """
-        # noinspection PyTypeChecker
-        return tgt.write('{}\n'.format(self.__password))
-
-    def connect(
-        self,
-        client,  # type: _type_ConnectSSH
-        hostname=None,  # type: typing.Optional[str]
-        port=22,  # type: int
-        log=True,  # type: bool
-    ):  # type: (...) -> None
-        """Connect SSH client object using credentials.
-
-        :param client: SSH Client (low level)
-        :type client: typing.Union[
-            paramiko.client.SSHClient,
-            paramiko.transport.Transport,
-        ]
-        :param hostname: remote hostname
-        :type hostname: str
-        :param port: remote ssh port
-        :type port: int
-        :param log: Log on generic connection failure
-        :type log: bool
-        :raises: paramiko.AuthenticationException
-        """
-        kwargs = {
-            'username': self.username,
-            'password': self.__password,
-            'key_filename': self.key_filename,
-            'passphrase': self.__passphrase,
-        }
-        if hostname is not None:
-            kwargs['hostname'] = hostname
-            kwargs['port'] = port
-
-        keys = [self.__key]
-        keys.extend([k for k in self.__keys if k != self.__key])
-
-        for key in keys:
-            kwargs['pkey'] = key
-            try:
-                client.connect(**kwargs)
-                if self.__key != key:
-                    self.__key = key
-                    logger.debug(
-                        'Main key has been updated, public key is: \n'
-                        '{}'.format(self.public_key))
-                return
-            except paramiko.PasswordRequiredException:
-                if self.__password is None:
-                    logger.exception('No password has been set!')
-                    raise
-                else:
-                    logger.critical(
-                        'Unexpected PasswordRequiredException, '
-                        'when password is set!')
-                    raise
-            except (paramiko.AuthenticationException,
-                    paramiko.BadHostKeyException):
-                continue
-        msg = 'Connection using stored authentication info failed!'
-        if log:
-            logger.exception(msg)
-        raise paramiko.AuthenticationException(msg)
-
-    def __hash__(self):
-        """Hash for usage as dict keys and comparison."""
-        return hash((
-            self.__class__,
-            self.username,
-            self.__password,
-            tuple(self.__keys),
-            (
-                tuple(self.key_filename)
-                if isinstance(self.key_filename, list)
-                else self.key_filename
-            ),
-            self.__passphrase
-        ))
-
-    def __eq__(self, other):  # type: (typing.Any) -> bool
-        """Comparison helper."""
-        return hash(self) == hash(other)
-
-    def __ne__(self, other):  # type: (typing.Any) -> bool
-        """Comparison helper."""
-        return not self.__eq__(other)
-
-    def __deepcopy__(self, memo):  # type: (typing.Any) -> SSHAuth
-        """Helper for copy.deepcopy."""
-        return self.__class__(
-            username=self.username,
-            password=self.__password,
-            key=self.__key,
-            keys=copy.deepcopy(self.__keys)
-        )
-
-    def __copy__(self):  # type: () -> SSHAuth
-        """Copy self."""
-        return self.__class__(
-            username=self.username,
-            password=self.__password,
-            key=self.__key,
-            keys=self.__keys
-        )
-
-    def __repr__(self):
-        """Representation for debug purposes."""
-        _key = (
-            None if self.__key is None else
-            '<private for pub: {}>'.format(self.public_key)
-        )
-        _keys = []
-        for k in self.__keys:
-            if k == self.__key:
-                continue
-            # noinspection PyTypeChecker
-            _keys.append(
-                '<private for pub: {}>'.format(
-                    self.__get_public_key(key=k)) if k is not None else None)
-
-        return (
-            '{cls}('
-            'username={self.username!r}, '
-            'password=<*masked*>, '
-            'key={key}, '
-            'keys={keys}, '
-            'key_filename={self.key_filename!r}, '
-            'passphrase=<*masked*>,'
-            ')'.format(
-                cls=self.__class__.__name__,
-                self=self,
-                key=_key,
-                keys=_keys)
-        )
-
-    def __str__(self):
-        """Representation for debug purposes."""
-        return (
-            '{cls} for {username}'.format(
-                cls=self.__class__.__name__,
-                username=self.username,
-            )
-        )
 
 
 class _MemorizedSSH(type):
@@ -348,7 +105,7 @@ class _MemorizedSSH(type):
         username=None,  # type: typing.Optional[str]
         password=None,  # type: typing.Optional[str]
         private_keys=None,  # type: typing.Optional[_type_RSAKeys]
-        auth=None,  # type: typing.Optional[SSHAuth]
+        auth=None,  # type: typing.Optional[ssh_auth.SSHAuth]
     ):  # type: (...) -> SSHClientBase
         """Main memorize method: check for cached instance and return it.
 
@@ -357,13 +114,13 @@ class _MemorizedSSH(type):
         :type username: str
         :type password: str
         :type private_keys: list
-        :type auth: SSHAuth
+        :type auth: ssh_auth.SSHAuth
         :rtype: SSHClient
         """
         if (host, port) in cls.__cache:
             key = host, port
             if auth is None:
-                auth = SSHAuth(
+                auth = ssh_auth.SSHAuth(
                     username=username,
                     password=password,
                     keys=private_keys
@@ -496,7 +253,7 @@ class SSHClientBase(BaseSSHClient):
         username=None,  # type: typing.Optional[str]
         password=None,  # type: typing.Optional[str]
         private_keys=None,  # type: typing.Optional[_type_RSAKeys]
-        auth=None,  # type: typing.Optional[SSHAuth]
+        auth=None,  # type: typing.Optional[ssh_auth.SSHAuth]
     ):
         """SSHClient helper.
 
@@ -511,7 +268,7 @@ class SSHClientBase(BaseSSHClient):
         :param private_keys: private keys for connection
         :type private_keys: typing.Optional[typing.Iterable[paramiko.RSAKey]]
         :param auth: credentials for connection
-        :type auth: typing.Optional[SSHAuth]
+        :type auth: typing.Optional[ssh_auth.SSHAuth]
 
         .. note:: auth has priority over username/password/private_keys
         """
@@ -526,7 +283,7 @@ class SSHClientBase(BaseSSHClient):
         self.__sftp = None
 
         if auth is None:
-            self.__auth = SSHAuth(
+            self.__auth = ssh_auth.SSHAuth(
                 username=username,
                 password=password,
                 keys=private_keys
@@ -556,7 +313,7 @@ class SSHClientBase(BaseSSHClient):
         return self.__logger
 
     @property
-    def auth(self):  # type: () -> SSHAuth
+    def auth(self):  # type: () -> ssh_auth.SSHAuth
         """Internal authorisation object.
 
         Attention: this public property is mainly for inheritance,
@@ -564,7 +321,7 @@ class SSHClientBase(BaseSSHClient):
         Calls outside SSHClient and child classes is sign of incorrect design.
         Change is completely disallowed.
 
-        :rtype: SSHAuth
+        :rtype: ssh_auth.SSHAuth
         """
         return self.__auth
 
@@ -1041,7 +798,7 @@ class SSHClientBase(BaseSSHClient):
         self,
         hostname,  # type: str
         command,  # type: str
-        auth=None,  # type: typing.Optional[SSHAuth]
+        auth=None,  # type: typing.Optional[ssh_auth.SSHAuth]
         target_port=22,  # type: int
         verbose=False,  # type: bool
         timeout=None,  # type: typing.Optional[int]
@@ -1055,7 +812,7 @@ class SSHClientBase(BaseSSHClient):
         :param command: Command for execution
         :type command: str
         :param auth: credentials for target machine
-        :type auth: typing.Optional[SSHAuth]
+        :type auth: typing.Optional[ssh_auth.SSHAuth]
         :param target_port: target port
         :type target_port: int
         :param verbose: Produce log.info records for command call and output
